move webview into custom-snippets for build
CI fails when I add the new file webview.ts,
so to work around that I'll just add the webview to snippets
